renaming store classes
