+++ lib/resources/__tests__/BaseTest.js
@@ -5,7 +5,7 @@ jest.dontMock("../../utilities/URLHelpers");
 
 import Constants from "../../Constants";
 
-describe("BaseActions", () => {
+describe("Base", () => {
   var actions, dispatcher;
 
   beforeEach(() => {
