+++ lib/stores/Orders.js
@@ -1,13 +1,13 @@
 "use strict";
 
 import APIDispatcher from "../dispatcher/APIDispatcher";
-import BaseStore     from "./BaseStore";
+import Base          from "./Base";
 import Constants     from "../Constants";
 
-class OrdersStore extends BaseStore {
+class Orders extends Base {
   constructor(dispatcher) {
     super(dispatcher, Constants.Actions.MERGE_ORDERS, Constants.Actions.CLEAR_ORDERS);
   }
 }
 
-export default new OrdersStore(APIDispatcher);
+export default new Orders(APIDispatcher);
