+++ lib/stores/Products.js
@@ -1,13 +1,13 @@
 "use strict";
 
 import APIDispatcher from "../dispatcher/APIDispatcher";
-import BaseStore     from "./BaseStore";
+import Base          from "./Base";
 import Constants     from "../Constants";
 
-class ProductsStore extends BaseStore {
+class Products extends Base {
   constructor(dispatcher) {
     super(dispatcher, Constants.Actions.MERGE_PRODUCTS, Constants.Actions.CLEAR_PRODUCTS);
   }
 }
 
-export default new ProductsStore(APIDispatcher);
+export default new Products(APIDispatcher);
