+++ lib/stores/Sessions.js
@@ -9,7 +9,7 @@ var accessToken = null;
 
 const { Store } = Flux;
 
-class SessionStore extends Store {
+class Sessions extends Store {
   getDomain() {
     return domain;
   }
@@ -31,4 +31,4 @@ class SessionStore extends Store {
   }
 }
 
-export default new SessionStore(APIDispatcher);
+export default new Sessions(APIDispatcher);
