+++ lib/stores/Shops.js
@@ -8,7 +8,7 @@ var shop = null;
 
 const { Store } = Flux;
 
-class ShopStore extends Store {
+class Shops extends Store {
   getCurrent() {
     return shop;
   }
@@ -30,4 +30,4 @@ class ShopStore extends Store {
   }
 }
 
-export default new ShopStore(APIDispatcher);
+export default new Shops(APIDispatcher);
