+++ lib/stores/__tests__/BaseTest.js
@@ -2,11 +2,11 @@
 
 jest.dontMock("../../dispatcher/APIDispatcher");
 jest.dontMock("flux/utils");
-jest.dontMock("../BaseStore");
+jest.dontMock("../Base");
 
 import Constants from "../../Constants";
 
-const BaseStore = require("../BaseStore");
+const BaseStore = require("../Base");
 
 class ThingsStore extends BaseStore {
   constructor(dispatcher) {
@@ -14,7 +14,7 @@ class ThingsStore extends BaseStore {
   }
 }
 
-describe("BaseStore", () => {
+describe("Base", () => {
   var store, dispatcher;
 
   beforeEach(() => {
