+++ lib/stores/__tests__/OrdersTest.js
@@ -1,15 +1,15 @@
 "use strict";
 
-jest.dontMock("../BaseStore");
-jest.dontMock("../OrdersStore");
+jest.dontMock("../Base");
+jest.dontMock("../Orders");
 
 import Constants from "../../Constants";
 
-describe("OrdersStore", () => {
+describe("Orders", () => {
   var store;
 
   beforeEach(() => {
-    store = require("../OrdersStore");
+    store = require("../Orders");
   });
 
   it("correctly defines mergeAction", () => {
