+++ lib/stores/__tests__/ProductsTest.js
@@ -1,15 +1,15 @@
 "use strict";
 
-jest.dontMock("../BaseStore");
-jest.dontMock("../ProductsStore");
+jest.dontMock("../Base");
+jest.dontMock("../Products");
 
 import Constants from "../../Constants";
 
-describe("ProductsStore", () => {
+describe("Products", () => {
   var store;
 
   beforeEach(() => {
-    store = require("../ProductsStore");
+    store = require("../Products");
   });
 
   it("correctly defines mergeAction", () => {
