+++ lib/stores/__tests__/SessionsTest.js
@@ -1,14 +1,14 @@
 "use strict";
 
 jest.dontMock("flux/utils");
-jest.dontMock("../SessionStore");
+jest.dontMock("../Sessions");
 
-describe("SessionStore", () => {
+describe("Session", () => {
   var SessionStore, APIDispatcher;
 
   beforeEach(() => {
     APIDispatcher = require("../../dispatcher/APIDispatcher");
-    SessionStore  = require("../SessionStore");
+    SessionStore  = require("../Sessions");
   });
 
   it("registers a callback with the dispatcher", () => {
