+++ lib/stores/__tests__/ShopsTest.js
@@ -2,15 +2,15 @@
 
 jest.dontMock("../../dispatcher/APIDispatcher");
 jest.dontMock("flux/utils");
-jest.dontMock("../ShopStore");
+jest.dontMock("../Shops");
 
 import Constants from "../../Constants";
 
-describe("ShopStore", () => {
+describe("Shop", () => {
   var store, dispatcher, changeEvent;
 
   beforeEach(() => {
-    store      = require("../ShopStore");
+    store      = require("../Shops");
     dispatcher = store.getDispatcher();
 
     changeEvent = jest.genMockFunction();
