+++ lib/utilities/WebAPI.js
@@ -1,7 +1,7 @@
 "use strict";
 
 import Constants    from "../Constants";
-import SessionStore from "../stores/SessionStore";
+import SessionStore from "../stores/Sessions";
 
 function _performRequest(method, url, options) {
   options = (options || {});
