+++ lib/utilities/__tests__/APIProxyTest.js
@@ -13,7 +13,7 @@ describe("APIProxy", () => {
     actions       = require("../../resources/Product");
     actions.fetch = jest.genMockFunction();
 
-    store = require("../../stores/ProductsStore");
+    store = require("../../stores/Products");
     api   = new APIProxy(actions, store);
   });
