+++ lib/utilities/__tests__/WebAPITest.js
@@ -8,7 +8,7 @@ describe("WebAPI", () => {
   var SessionStore, WebAPI;
 
   beforeEach(() => {
-    SessionStore = require("../../stores/SessionStore");
+    SessionStore = require("../../stores/Sessions");
     WebAPI       = require("../WebAPI");
 
     SessionStore.getDomain.mockReturnValue("example.myshopify.com");
